Extract prometheus.go from infra_test.go
Extract several utility functions into libinfra
