+++ tests/libinfra/BUILD.bazel
@@ -4,17 +4,28 @@ go_library(
     name = "go_default_library",
     srcs = [
         "certificates.go",
+        "http.go",
+        "leader.go",
         "metrics.go",
+        "network.go",
     ],
     importpath = "kubevirt.io/kubevirt/tests/libinfra",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/certificates/triple/cert:go_default_library",
         "//pkg/downwardmetrics/vhostmd/api:go_default_library",
+        "//pkg/virt-controller/leaderelectionconfig:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//tests/console:go_default_library",
+        "//tests/flags:go_default_library",
+        "//tests/framework/kubevirt:go_default_library",
+        "//tests/libnet:go_default_library",
+        "//tests/util:go_default_library",
         "//vendor/github.com/google/goexpect:go_default_library",
         "//vendor/github.com/onsi/ginkgo/v2:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
+        "//vendor/k8s.io/api/core/v1:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
+        "//vendor/k8s.io/client-go/tools/leaderelection/resourcelock:go_default_library",
     ],
 )
